자잘한 수정

+++ src/main/java/com/ss/heartlinkapi/admin/controller/AdminCoupleController.java
@@ -4,6 +4,8 @@ import com.ss.heartlinkapi.admin.service.AdminCoupleService;
 import com.ss.heartlinkapi.linkmatch.entity.LinkMatchEntity;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -24,20 +26,36 @@ public class AdminCoupleController {
 
     // 매치 질문 등록
     @PostMapping("/questions")
-    public void addMatchQuestion(@RequestBody LinkMatchEntity questionText) {
-        adminCoupleService.addMatchQuestion(questionText);
+    public ResponseEntity<?> addMatchQuestion(@RequestBody LinkMatchEntity questionText) {
+        LinkMatchEntity result = adminCoupleService.addMatchQuestion(questionText);
+        if(result != null) {
+            return ResponseEntity.ok(result);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     // 매치 질문 수정
     @PutMapping("/questions/{questionId}/update")
-    public void updateMatchQuestion(@PathVariable Long questionId, @RequestBody LinkMatchEntity questionText) {
-        adminCoupleService.updateMatchQuestion(questionId, questionText);
+    public ResponseEntity<?> updateMatchQuestion(@PathVariable Long questionId, @RequestBody LinkMatchEntity questionText) {
+        LinkMatchEntity result = adminCoupleService.updateMatchQuestion(questionId, questionText);
+        if(result != null) {
+            return ResponseEntity.ok(result);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     // 매치 질문 삭제
     @DeleteMapping("/questions/{questionId}/delete")
-    public void deleteMatchQuestion(@PathVariable Long questionId) {
-        adminCoupleService.deleteMatchQuestion(questionId);
+    public ResponseEntity<?> deleteMatchQuestion(@PathVariable Long questionId) {
+        LinkMatchEntity question = adminCoupleService.findById(questionId);
+        if(question != null) {
+            adminCoupleService.deleteMatchQuestion(questionId);
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
 }
